Feature/2 d world

+++ GameServer/Physics/Physic.cs
@@ -8,7 +8,7 @@ namespace GameServer.Physics
 {
     public class Physic
     {
-        public static Vector2 GetLeftParallelVectorToIntersectionNormal(Vector2 movementVector, float intersectionDistance, Vector2 intersectionNormal)
+        public static Vector2 GetLeftParallelVectorToIntersectionNormal(Vector2 movementVector, double intersectionDistance, Vector2 intersectionNormal)
         {
             var movementVectorLength = movementVector.Length();
             var realMovementDirectionVector = Vector2.Normalize(movementVector) * intersectionDistance;
@@ -32,16 +32,16 @@ namespace GameServer.Physics
             return new Vector2(x, y);
         }
 
-        public static Vector2 RotateVector(Vector2 v, float angle)
+        public static Vector2 RotateVector(Vector2 v, double angle)
         {
-            var ca = (float)Math.Cos(angle);
-            var sa = (float)Math.Sin(angle);
+            var ca = Math.Cos(angle);
+            var sa = Math.Sin(angle);
             return new Vector2(ca * v.X - sa * v.Y, sa * v.X + ca * v.Y);
         }
 
-        public static Vector2 AngleToVector(float angle)
+        public static Vector2 AngleToVector(double angle)
         {
-            return new Vector2((float)Math.Sin(angle), (float)Math.Cos(angle));
+            return new Vector2(Math.Sin(angle), Math.Cos(angle));
         }
 
         public static float VectorToAngle(Vector2 vector)
